Mavlink socket communication running and tested
